fix(KEN-606): the invented pin gives way on both sides [no-changelog]
Address PR review. Three defects the first pass introduced.

The exemption is narrowed by relation, not wholesale. Only the sets a
target's own installations name as carrying it are unexempted; one reached
through a dependency parent stays exempt, because it owns that parent's
revision rather than this target's. Pinned, the parent carried its commit
onto a target reading fresh and the update reported a conflict instead of
happening.

Where a set is held is said by the members whose revision it still
carries. An installation the person also declared reads its own
declaration, and a single-package update moves that commit off the set's;
counted as the set's evidence, the members read as disagreeing, the set
was placeable at no commit at all, and its undeclared members came current
on the next update of anything else.

And the rule that only a revision this pass invented gives way now holds
on the member side too. A declaration keeps the revision the person chose
for it, and that is what a set's revision is weighed against — never the
pin invented to hold it still, which read as agreement where the two
matched and named a commit kendex made up where they did not.

The set tests moved to their own file: the old one crossed the size
ratchet's threshold for a test, and set membership is where it splits.
